@@ -5,6 +5,7 @@ use std::sync::{Arc, Mutex};
 use std::time::Duration;
 use tokio::time::Instant;
 
+use crate::async_impl::body::ResponseBody;
 use crate::error::{BoxError, Error, Kind};
 use crate::Body;
 use bytes::Buf;
@@ -13,7 +14,6 @@ use h3::client::SendRequest;
 use h3_quinn::{Connection, OpenStreams};
 use http::uri::{Authority, Scheme};
 use http::{Request, Response, Uri};
-use hyper::Body as HyperBody;
 use log::trace;
 
 pub(super) type Key = (Scheme, Authority);
@@ -37,7 +37,7 @@ impl Pool {
     pub fn connecting(&self, key: Key) -> Result<(), BoxError> {
         let mut inner = self.inner.lock().unwrap();
         if !inner.connecting.insert(key.clone()) {
-            return Err(format!("HTTP/3 connecting already in progress for {:?}", key).into());
+            return Err(format!("HTTP/3 connecting already in progress for {key:?}").into());
         }
         return Ok(());
     }
@@ -77,7 +77,7 @@ impl Pool {
         let (close_tx, close_rx) = std::sync::mpsc::channel();
         tokio::spawn(async move {
             if let Err(e) = future::poll_fn(|cx| driver.poll_close(cx)).await {
-                trace!("poll_close returned error {:?}", e);
+                trace!("poll_close returned error {e:?}");
                 close_tx.send(e).ok();
             }
         });
@@ -105,7 +105,7 @@ struct PoolInner {
 impl PoolInner {
     fn insert(&mut self, key: Key, conn: PoolConnection) {
         if self.idle_conns.contains_key(&key) {
-            trace!("connection already exists for key {:?}", key);
+            trace!("connection already exists for key {key:?}");
         }
 
         self.idle_conns.insert(key, conn);
@@ -125,9 +125,20 @@ impl PoolClient {
     pub async fn send_request(
         &mut self,
         req: Request<Body>,
-    ) -> Result<Response<HyperBody>, BoxError> {
+    ) -> Result<Response<ResponseBody>, BoxError> {
+        use http_body_util::{BodyExt, Full};
+        use hyper::body::Body as _;
+
         let (head, req_body) = req.into_parts();
-        let req = Request::from_parts(head, ());
+        let mut req = Request::from_parts(head, ());
+
+        if let Some(n) = req_body.size_hint().exact() {
+            if n > 0 {
+                req.headers_mut()
+                    .insert(http::header::CONTENT_LENGTH, n.into());
+            }
+        }
+
         let mut stream = self.inner.send_request(req).await?;
 
         match req_body.as_bytes() {
@@ -146,7 +157,11 @@ impl PoolClient {
             resp_body.extend(chunk.chunk())
         }
 
-        Ok(resp.map(|_| HyperBody::from(resp_body)))
+        let resp_body = Full::new(resp_body.into())
+            .map_err(|never| match never {})
+            .boxed();
+
+        Ok(resp.map(|_| resp_body))
     }
 }
 
