@@ -1,9 +1,9 @@
 //! DNS resolution
 
-pub use resolve::{Addrs, Resolve, Resolving};
+pub use resolve::{Addrs, Name, Resolve, Resolving};
 pub(crate) use resolve::{DnsResolverWithOverrides, DynResolver};
 
 pub(crate) mod gai;
+#[cfg(feature = "hickory-dns")]
+pub(crate) mod hickory;
 pub(crate) mod resolve;
-#[cfg(feature = "trust-dns")]
-pub(crate) mod trust_dns;
